Add storefinder
